Update javascrip to interacte with the html
Functions:

-getComputerChoice
Randomly chose a number 1 to 3 and return a string
Use dom to update hand image of computer

-playGame
--playRound
Uses humanChoice and getComputerChoice to print
messages and updates scores.

Have an alert if player doesnt chose a hand.

-reset
When the game hits 5 on a score reset definitions

DOM events
-Choose player hand and highlight the one selected
-Fight button to activate functions to play and
when scrores reach 5 start a reset.
-Computer image updates according is hand.